Dead end marker implemented (using DFS logic, not actual dead-end detection search) + there's a bug regarding the data types of the coordinates

+++ Explorer/src/main/java/nl/maastrichtuniversity/dke/explorer/Logic/Objects/Object.java
@@ -5,7 +5,7 @@ import java.awt.image.BufferedImage;
 
 public class Object {
 
-    public int x, y;
+    public double x, y;
     public int markerType;
     public boolean interact;
     public BufferedImage image;
@@ -15,16 +15,16 @@ public class Object {
         this.image = image;
     }
 
-    public void setCoord(int x, int y) {
+    public void setCoord(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
     public void setMarkerType(int mT) { this.markerType = mT; }
 
-    public int getX() { return this.x; }
+    public double getX() { return this.x; }
 
-    public int getY() { return this.y; }
+    public double getY() { return this.y; }
 
     public int getMarkerType() { return this.markerType; }
